Rename opengl/vulkan.h to _private.h
