+++ engine/graphics/src/vulkan/macosx/graphics_vulkan_macosx_surface.mm
@@ -1,5 +1,5 @@
 
-#include "../graphics_vulkan.h"
+#include "../graphics_vulkan_private.h"
 
 #include <graphics/glfw/glfw_native.h>
 #include <vulkan/vulkan_macos.h>
