+++ engine/graphics/src/vulkan/win32/graphics_vulkan_win32_surface.cpp
@@ -1,5 +1,5 @@
 
-#include "../graphics_vulkan.h"
+#include "../graphics_vulkan_private.h"
 
 #include <string.h>
